minor fix - removed warning on 64-bit platforms when converting between pointer  and integers.
git-svn-id: http://dyncall.org/svn/dyncall/trunk/dyncall@894 c71cb566-5a31-4692-9165-cf76bdc42cab

+++ test/suite/main.c
@@ -69,7 +69,7 @@ void init()
     valueInt[i]      = (DCint)     (i);
     valueLongLong[i] = (DClonglong)(i);
     valueDouble[i]   = (DCdouble)  (i);
-    valuePointer[i]  = (DCpointer) (i);
+    valuePointer[i]  = (DCpointer) (ptrdiff_t) (i);
     valueFloat[i]    = (DCfloat)   (i);
   } 
 }
